pyproject.toml
[project]
name = "pyreview"
version = "0.1.0"
description = "Multi-agent AI code review system powered by Claude"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.43.0",
    "typer[all]>=0.15.0",
    "rich>=14.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "websockets>=14.0",
    "python-multipart>=0.0.18",
    "PyGithub>=2.5.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "unidiff>=0.7.5",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "respx>=0.22.0",
]

[project.scripts]
pyreview = "pyreview.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyreview"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
